BUGFIX: correct position of tile objects
On an object layer, a tile object's position is the bottom-leftmost part of the object. This is because, like all other objects, you can stretch and skew them. However, this doesn't mesh with sprite positioning, because sprites are positioned based on their center-point.

In order to deal with this, when setting the position of a tile object, we take the half width and half height in order to gain the true center and use that adjusted position as the sprite's position.